+++ src/Devices/GoodDevices.php
@@ -113,8 +113,30 @@ class GoodDevices
      * LAST-UPDATED: MARCH 2017.
      */
     const DEVICES = [
-        '24/7.0; 480dpi; 1080x1920; samsung; SM-A520F; a5y17lte; samsungexynos7880',
-        '22/5.1; 480dpi; 1080x1920; Meizu; m3 note; m3note; mt6755',
+        // SAMSUNNG GALAXY S9 (EUR AND USA MODELS).
+        '28/9.0; 560dpi; 1440x2960; samsung; SM-G960F; starlte; samsungexynos9810',
+        '26/8.0; 560dpi; 1440x2960; samsung; SM-G960F; starlte; samsungexynos9810',
+        '28/9.0; 560dpi; 1440x2960; samsung; SM-G960U; starlte; samsungexynos9810',
+        '26/8.0; 560dpi; 1440x2960; samsung; SM-G960U; starlte; samsungexynos9810',
+
+        // SAMSUNNG GALAXY S8 (USA AND EUR MODELS).
+        '28/9.0; 480dpi; 1440x2960; samsung; SM-G950A; dreamlte; samsungexynos8895',
+        '27/8.1; 480dpi; 1440x2960; samsung; SM-G950A; dreamlte; samsungexynos8895',
+        '26/8.0; 480dpi; 1440x2960; samsung; SM-G950A; dreamlte; samsungexynos8895',
+        '24/7.0; 480dpi; 1440x2960; samsung; SM-G950A; dreamlte; samsungexynos8895',
+        '28/9.0; 480dpi; 1440x2960; samsung; SM-G950F; dreamlte; samsungexynos8895',
+        '27/8.1; 480dpi; 1440x2960; samsung; SM-G950F; dreamlte; samsungexynos8895',
+        '26/8.0; 480dpi; 1440x2960; samsung; SM-G950F; dreamlte; samsungexynos8895',
+        '24/7.0; 480dpi; 1440x2960; samsung; SM-G950F; dreamlte; samsungexynos8895',
+
+        // SAMSUNNG GALAXY S7 (EUR AND USA MODELS).
+        '26/8.0; 560dpi; 1440x2560; samsung; SM-G930F; herolte; samsungexynos8890',
+        '25/7.1; 560dpi; 1440x2560; samsung; SM-G930F; herolte; samsungexynos8890',
+        '23/6.0; 560dpi; 1440x2560; samsung; SM-G930F; herolte; samsungexynos8890',
+        '26/8.0; 560dpi; 1440x2560; samsung; SM-G930U; herolte; samsungexynos8890',
+        '25/7.1; 560dpi; 1440x2560; samsung; SM-G930U; herolte; samsungexynos8890',
+        '23/6.0; 560dpi; 1440x2560; samsung; SM-G930U; herolte; samsungexynos8890',
+
         '24/7.0; 640dpi; 1440x2560; samsung; SM-G925F; zerolte; samsungexynos7420',
         '23/6.0.1; 480dpi; 1080x1920; Xiaomi; Redmi Note 3; kenzo; qcom',
         '24/7.0; 480dpi; 1080x2076; samsung; SM-G950F; dreamlte; samsungexynos8895',
@@ -170,6 +192,34 @@ class GoodDevices
         '26/8.0.0; 480dpi; 1080x1920; samsung; SM-G930F; herolte; samsungexynos8890',
         '27/8.1.0; 440dpi; 1080x2030; Xiaomi/xiaomi; Redmi 5 Plus; vince; qcom',
         '28/9; 560dpi; 1440x2792; samsung; SM-G965N; star2lteks; samsungexynos9810',
+        '25/7.1.1; 440dpi; 1080x1920; Xiaomi; Mi Note 3; jason; qcom',
+        '23/6.0.1; 480dpi; 1080x1920; Xiaomi; Redmi Note 3; kenzo; qcom',
+        '23/6.0; 480dpi; 1080x1920; Xiaomi; Redmi Note 4; nikel; mt6797',
+        '24/7.0; 480dpi; 1080x1920; Xiaomi/xiaomi; Redmi Note 4; mido; qcom',
+        '27/8.1.0; 440dpi; 1080x2030; Xiaomi/xiaomi; Redmi Note 5; whyred; qcom',
+        '23/6.0.1; 480dpi; 1080x1920; Xiaomi; Redmi 4; markw; qcom',
+        '25/7.1.2; 440dpi; 1080x2030; Xiaomi/xiaomi; Redmi 5 Plus; vince; qcom',
+        '26/8.0.0; 480dpi; 1080x1920; Xiaomi; MI 5; gemini; qcom',
+        '26/8.0.0; 480dpi; 1080x1920; Xiaomi; MI 6; sagit; qcom',
+        '25/7.1.1; 440dpi; 1080x1920; Xiaomi; MI MAX 2; oxygen; qcom',
+        '24/7.0; 480dpi; 1080x1920; Xiaomi; MI 5s; capricorn; qcom',
+        '26/8.0.0; 480dpi; 1080x1920; samsung; SM-A520F; a5y17lte; samsungexynos7880',
+        '26/8.0.0; 480dpi; 1080x2076; samsung; SM-G950F; dreamlte; samsungexynos8895',
+        '26/8.0.0; 640dpi; 1440x2768; samsung; SM-G950F; dreamlte; samsungexynos8895',
+        '26/8.0.0; 420dpi; 1080x2094; samsung; SM-G955F; dream2lte; samsungexynos8895',
+        '26/8.0.0; 480dpi; 1080x1920; samsung; SM-G935F; hero2lte; samsungexynos8890',
+        '26/8.0.0; 480dpi; 1080x2076; samsung; SM-A530F; jackpotlte; samsungexynos7885',
+        '24/7.0; 640dpi; 1440x2560; samsung; SM-G925F; zerolte; samsungexynos7420',
+        '26/8.0.0; 420dpi; 1080x1920; samsung; SM-A720F; a7y17lte; samsungexynos7880',
+        '24/7.0; 640dpi; 1440x2560; samsung; SM-G920F; zeroflte; samsungexynos7420',
+        '24/7.0; 420dpi; 1080x1920; samsung; SM-J730FM; j7y17lte; samsungexynos7870',
+        '26/8.0.0; 420dpi; 1080x2094; samsung; SM-A730F; jackpot2lte; samsungexynos7885',
+        '26/8.0.0; 420dpi; 1080x2094; samsung; SM-A605FN; a6plte; qcom',
+        '27/8.1.0; 480dpi; 1080x2280; HUAWEI/HONOR; COL-L29; HWCOL; kirin970',
+        '26/8.0.0; 480dpi; 1080x2032; HUAWEI; FIG-LX1; HWFIG-H; hi6250',
+        '27/8.1.0; 480dpi; 1080x2150; HUAWEI/HONOR; COL-L29; HWCOL; kirin970',
+        '26/8.0.0; 480dpi; 1080x2038; HUAWEI/HONOR; BND-L21; HWBND-H; hi6250',
+        '23/6.0.1; 420dpi; 1080x1920; LeMobile/LeEco; Le X527; le_s2_ww; qcom',
     ];
 
     const VP9_COMPATIBLE_DEVICES = [